tasks: make `setup:install:homebrew` work correctly
Given the goal is to make everything “just work” in bash, go ahead and
put things in `.bashrc` for now. This is *not* especially robust: it
does not correctly handle the login/non-login shell distinction, and I
will need to change it in the future. It does, however, unblock things
for the moment!